+++ api/fs/fs.h
@@ -62,6 +62,11 @@ struct TempFileResult {
     string path;
 };
 
+struct TempDirectoryResult {
+    errors::StatusCode status = errors::NE_ST_OK;
+    string path;
+};
+
 fs::FileReaderResult readFile(const string &filename, const fs::FileReaderOptions &fileReaderOptions = {});
 bool writeFile(const fs::FileWriterOptions &fileWriterOptions);
 string getDirectoryName(const string &filename);
@@ -73,12 +78,14 @@ bool removeWatcher(long watcherId);
 fs::FileStats getStats(const string &path);
 fs::DirReaderResult readDirectory(const string &path, bool recursive = false);
 fs::TempFileResult createTempFile(const string &prefix = "", const string &extension = "");
+fs::TempDirectoryResult createTempDirectory(const string &prefix = "");
 string applyPathConstants(const string &path);
 
 namespace controllers {
 
 json createDirectory(const json &input);
 json createTempFile(const json &input);
+json createTempDirectory(const json &input);
 json remove(const json &input);
 json writeFile(const json &input);
 json writeBinaryFile(const json &input);